Rename `attrutil` to `attributes`.
