+++ src/rinq/internal/remotesession/catalog.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/rinq/rinq-go/src/rinq"
 	"github.com/rinq/rinq-go/src/rinq/ident"
-	"github.com/rinq/rinq-go/src/rinq/internal/attrutil"
+	"github.com/rinq/rinq-go/src/rinq/internal/attributes"
 	revisionpkg "github.com/rinq/rinq-go/src/rinq/internal/revision"
 	"github.com/rinq/rinq-go/src/rinq/internal/x/syncx"
 )
@@ -77,7 +77,7 @@ func (c *catalog) Fetch(
 	rev ident.Revision,
 	ns string,
 	keys ...string,
-) (attrutil.List, error) {
+) (attributes.List, error) {
 	solvedAttrs, unsolvedKeys, err := c.fetchLocal(rev, ns, keys)
 	if err != nil {
 		return nil, err
@@ -151,7 +151,7 @@ func (c *catalog) TryUpdate(
 	ctx context.Context,
 	rev ident.Revision,
 	ns string,
-	attrs attrutil.List,
+	attrs attributes.List,
 ) (rinq.Revision, error) {
 	unlock := syncx.RLock(&c.mutex)
 	defer unlock()
@@ -166,7 +166,7 @@ func (c *catalog) TryUpdate(
 		return nil, rinq.StaleUpdateError{Ref: ref}
 	}
 
-	updateAttrs := make(attrutil.List, 0, len(attrs))
+	updateAttrs := make(attributes.List, 0, len(attrs))
 
 	cache := c.cache[ns]
 
@@ -318,7 +318,7 @@ func (c *catalog) fetchLocal(
 	ns string,
 	keys []string,
 ) (
-	solved attrutil.List,
+	solved attributes.List,
 	unsolved []string,
 	err error,
 ) {
@@ -326,7 +326,7 @@ func (c *catalog) fetchLocal(
 	defer c.mutex.RUnlock()
 
 	count := len(keys)
-	solved = make(attrutil.List, 0, count)
+	solved = make(attributes.List, 0, count)
 	unsolved = make([]string, 0, count)
 
 	cache := c.cache[ns]
